Have KubeCluster accept only a Pod template
- Have a function, make_pod_spec that accepts a bunch of stuff
  and gives back a pod template
- This allows us to add creating pod templates from YAML, JSON,
  etc in the future
