+++ daskernetes/tests/test_core.py
@@ -4,12 +4,13 @@ from time import sleep, time
 
 import pytest
 from daskernetes import KubeCluster
+from daskernetes.objects import make_pod_spec
 from dask.distributed import Client
 from distributed.utils_test import loop, inc
 
 
-def test_basic(loop):
-    with KubeCluster(loop=loop) as cluster:
+def test_basic(image_name, loop):
+    with KubeCluster(make_pod_spec(image_name), loop=loop) as cluster:
         cluster.scale_up(2)
         with Client(cluster) as client:
             future = client.submit(inc, 10)
@@ -26,8 +27,8 @@ def test_basic(loop):
             assert all(client.has_what().values())
 
 
-def test_logs(loop):
-    with KubeCluster(loop=loop) as cluster:
+def test_logs(image_name, loop):
+    with KubeCluster(make_pod_spec(image_name), loop=loop) as cluster:
         cluster.scale_up(2)
 
         start = time()
@@ -40,9 +41,9 @@ def test_logs(loop):
         assert 'distributed.worker' in logs
 
 
-def test_ipython_display(loop):
+def test_ipython_display(image_name, loop):
     ipywidgets = pytest.importorskip('ipywidgets')
-    with KubeCluster(loop=loop) as cluster:
+    with KubeCluster(make_pod_spec(image_name), loop=loop) as cluster:
         cluster.scale_up(1)
         cluster._ipython_display_()
         box = cluster._cached_widget
@@ -58,16 +59,16 @@ def test_ipython_display(loop):
             sleep(0.5)
 
 
-def test_namespace(loop):
-    with KubeCluster(loop=loop) as cluster:
+def test_namespace(image_name, loop):
+    with KubeCluster(make_pod_spec(image_name), loop=loop) as cluster:
         assert 'dask' in cluster.name
         assert getpass.getuser() in cluster.name
-        with KubeCluster(loop=loop, port=0) as cluster2:
+        with KubeCluster(make_pod_spec(image_name), loop=loop) as cluster2:
             assert cluster.name != cluster2.name
 
 
-def test_adapt(loop):
-    with KubeCluster(loop=loop) as cluster:
+def test_adapt(image_name, loop):
+    with KubeCluster(make_pod_spec(image_name), loop=loop) as cluster:
         cluster.adapt()
         with Client(cluster) as client:
             future = client.submit(inc, 10)
@@ -80,8 +81,8 @@ def test_adapt(loop):
             assert time() < start + 10
 
 
-def test_env(loop):
-    with KubeCluster(loop=loop, env={'ABC': 'DEF'}) as cluster:
+def test_env(image_name, loop):
+    with KubeCluster(make_pod_spec(image_name, env={'ABC': 'DEF'}), loop=loop) as cluster:
         cluster.scale_up(1)
         with Client(cluster) as client:
             while not cluster.scheduler.workers:
